requirements.txt
flask
kiteconnect
python-dotenv
requests
mcp
fastmcp
